Made significant improvements to the `IMapViewProvider` hierarchy.
Add the `EchoAndroidView` composable. It is just a wrapper around an
`AndroidView` but it improves
readability.

Change the signature of the `update` method in `IMapViewProvider` to take advantage
of the way `AndroidView`s work (the `factory` function is only called
when the view is created, and the `update` function is called both when
the view is created and anytime the `AndroidView` is recomposed).

Adapt the `update` function and the constructor in `OsmdroidMapViewProvider` to comply with interface changes.

Improve the `MapDrawer` signature to to take a context instead of a
provider to facilitate usage.

Add documentation for relevant classes/composables.