Fixed navigation tests
This was more tricky than expected. To make the navigation tests work with hilt, I had to create a compose test rule based of the MainActivity. Then we can reset the content of the activity with whatever we want to test.

I spitted the tests for the navigation into to tests classes, NavigationActionsTest which test the NavigationActions and NavigationTest which tests the AppNavigationHost. In the NavigationTest I simply test that it is possible to navigate to all the routes.